+++ sw.js
@@ -13,7 +13,7 @@
  */
 'use strict';
 
-var VERSION = 'v65';
+var VERSION = 'v66';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -42,6 +42,7 @@ var SHELL_ASSETS = [
   './js/saved-state.js',
   './js/scene-codec.js',
   './js/multi-camera.js',
+  './js/map-comparison.js',
   './js/context-layers.js',
   './js/tropical-cyclones.js',
   './js/flood-outlooks.js',